rename guns.py to weapons.py
