+++ siegeapi/constants/__init__.py
@@ -2,4 +2,4 @@ from .stat_names import *
 from .seasons import seasons
 from .operators import operator_dict
 from .app_ids import SIEGE_APP_IDS
-from .guns import WEAPONS_DICT
+from .weapons import WEAPONS_DICT
