+++ siegeapi/constants/weapons.py
